requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
openai
anthropic
langchain
pypdf
python-multipart
langchain-community
bs4
arxiv
pymupdf